Add files via upload
This is a spam detection task done on twitter spam data; our aim is to separate the spam tweets from other tweets based on different categories within the dataset. Solving this problem is important in reducing spams and fake news from the social sites especially twitter. Whenever somebody will tweet a message containing spam, it will be marked as spam so that other people be aware of the spam before doing any activity on that tweet. This concept is also applicable on the emails and other social media messages/posts.